fix numpy dtype

+++ ada_endpoint.py
@@ -64,7 +64,7 @@ class AdaEndPoint:
         target_x = data["target_x"]
         target_y = data["target_x"]
         
-        img = np.asarray(data["image"])
+        img = np.asarray(data["image"], dtype=np.uint8)
         
         input_point = np.array([[target_x, target_y]])
         input_label = np.array([1])
@@ -114,7 +114,7 @@ class AdaEndPoint:
         #images = data["images"]
         images = []
         for img in data["images"]:
-            images.append(np.asarray(img))
+            images.append(np.asarray(img, dtype=np.uint8))
         text = data["text"]
     
         #org clip_image = self.clip_preprocess(images).unsqueeze(0).to(self.device)
